requirements.txt
kubernetes
openvino-python
pyusb
pyudev